Expose internal api on /internal/openapi.json
