+++ backend/src/test/java/com/redhat/cloud/notifications/openapi/OpenApiTest.java
@@ -1,6 +1,7 @@
 package com.redhat.cloud.notifications.openapi;
 
 import com.redhat.cloud.notifications.TestConstants;
+import com.redhat.cloud.notifications.oapi.OApiFilter;
 import com.redhat.cloud.notifications.routers.OApiService;
 import com.reprezen.kaizen.oasparser.OpenApi3Parser;
 import com.reprezen.kaizen.oasparser.model3.OpenApi3;
@@ -38,8 +39,11 @@ public class OpenApiTest {
     @TestHTTPResource(TestConstants.API_INTEGRATIONS_V_1 + "/openapi.json")
     URL iUrl1;
 
-    @TestHTTPResource("/api/" + OApiService.PRIVATE + "/v1.0/openapi.json")
+    @TestHTTPResource("/api/" + OApiFilter.PRIVATE + "/v1.0/openapi.json")
     URL privateUrl;
+
+    @TestHTTPResource("/" + OApiFilter.INTERNAL + "/openapi.json")
+    URL internalUrl;
 
     @TestHTTPResource("/api/doesNotExist/v1.0/openapi.json")
     URL badUrl;
@@ -47,7 +51,7 @@ public class OpenApiTest {
     @Test
     public void validateOpenApi() throws Exception {
 
-        URL[] urls = {nUrl, iUrl, nUrl1, iUrl1, privateUrl};
+        URL[] urls = {nUrl, iUrl, nUrl1, iUrl1, privateUrl, internalUrl};
 
         for (URL url : urls) {
             OpenApi3 model = new OpenApi3Parser().parse(url, true);
@@ -83,6 +87,10 @@ public class OpenApiTest {
 
         try (InputStream in = privateUrl.openStream()) {
             Files.copy(in, Paths.get("./target/openapi.private.json"), StandardCopyOption.REPLACE_EXISTING);
+        }
+
+        try (InputStream in = internalUrl.openStream()) {
+            Files.copy(in, Paths.get("./target/openapi.internal.json"), StandardCopyOption.REPLACE_EXISTING);
         }
     }
 }
